Add:
- A conditional check for Go and Rust toolchain updates, so no prompt saying "Skipping updates..." is displayed when running 'ropa update --full' and one or both of these toolchains aren't installed.

Change:
- Shorten the print function names.
- Fix typos.
- Change code section headers (comments).
- Remove some print spam, and reply more on the output of the package managers.
- Code fixes.
- Update docs.

Remove:
- Unnecessary comments that were duplicating the code.
- Experimental message for Go toolchain update function.

+++ ropa.d/go-package-update.sh
@@ -5,8 +5,6 @@
 # Copyright   : (c) 2025, Gergely Szabo
 # License     : MIT
 #
-# THIS FEATURE IS EXPERIMENTAL AND MAY NOT WORK AS EXPECTED!
-#
 # The script currently only supports non-root installations, for example: ~/go
 # or ~/.local/share/go.
 #
@@ -14,48 +12,47 @@
 #   ropa update|up --go|-g
 
 go_package_update() {
-  if command -v go &>/dev/null; then
+  # Fail silenty if called via 'ropa update --full'.
+  if [[ -z "$(command -v go)" && "$option" == "--full" ]] || \
+  [[ -z "$(command -v go)" && "$option" == "-f" ]]; then
+    return 0
+  fi
 
-    # Skip updates if Go is installed to /usr/local/ (officially recommended).
+  if command -v go &>/dev/null; then
     if [[ "$(command -v go)" == "/usr/local/go/bin/go" ]]; then
-      print_warning "Detected system-wide installation of Go toolchain. These are not supported right now, skipping updates."
+      print_warn "Detected system-wide installation of Go toolchain. These are not supported right now, skipping updates."
       return 0
     fi
 
-    print_action "Searching for Go toolchain updates..."
-
-    # Initialize the Go environment variables.
     GO_VERSION="$(go version | awk '{print $3}' | sed 's/go//')"
     GO_INSTALL_DIR="$(command -v go | sed 's|/go.*$||')"
     GO_LATEST_VER="$(curl -s https://go.dev/VERSION?m=text | grep -oP 'go\K[0-9.]+')"
 
-    # Check for updates & download the latest Go version.
+    print_step "Searching for Go toolchain updates..."
     if [[ "$GO_VERSION" == "$GO_LATEST_VER" ]]; then
-      print_success "No available Go toolchain updates."
+      print_info "No available Go toolchain updates."
     else
-      print_action "Updating Go toolchain..."
+      print_step "Updating Go toolchain..."
       wget "https://go.dev/dl/go$GO_LATEST_VER.linux-amd64.tar.gz" -O /tmp/go.tar.gz
 
       if [[ $? != 0 ]]; then
-        print_error "Failed to download the latest Go version. Please check the output for details."
+        print_err "Failed to download the latest Go version. Please check the output for details."
         return 1
       fi
 
-      # Remove current installation before installing new version.
+      # Remove current installation before installing the new version.
       rm -rf "$GO_INSTALL_DIR/go" &>/dev/null
-
-      # Install the new Go version.
       tar -C "$GO_INSTALL_DIR" -xzf /tmp/go.tar.gz
 
       if [[ $? == 0 ]]; then
-        print_success "Go toolchain has been updated."
+        print_info "Go toolchain has been updated."
       else
-        print_error "Failed to update Go. Please check the output for details."
+        print_err "Failed to update Go. Please check the output for details."
         return 1
       fi
     fi
   else
-    print_warning "Go toolchain is not installed. Skipping updates."
+    print_warn "Go toolchain is not installed. Skipping updates."
     return 0
   fi
